project/app/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
sqlalchemy
asyncpg
passlib
jose
python-jose
python-multipart
bcrypt
python-dotenv
aiofiles
pytz
alembic